Refactor NegationTwoPhaseIterator and expand tests
Extract shared scratch-fill/andNot/clear logic from intoBitSet() and
applyMask() into a private subtractExcluded() helper, eliminating the
duplication between the two methods.

Rename TestNegationIterator to TestNegationTwoPhaseIterator and expand
coverage: randomised correctness test for the TwoPhaseIterator excluded
path (including approximation false positives), direct tests for
intoBitSet() and applyMask(), docIDRunEnd() edge cases (NO-block with
fresh iterator, excl at current doc), all-excluded and none-excluded
edge cases, and a test verifying the scoring fallback to ReqExclBulkScorer.
